(SERVER-151) Use service context instead of app in jruby-testutils
The original versions of the jruby testutils that allowed interacting
with the environment cache were written to use a TK app as the
input.  This commit refactors them to use the `JRubyPuppetService`
service context as their input instead; this will be necessary
in order to wire up the HTTP endpoint, since we will only have
access to the service context and not to the entire TK app.