(SERVER-325) Re-organize code around flushing a JRuby instance
This commit simply makes a small reorganization to the code that
flushes an individual JRuby instance from the pool during
`flush-pool!`.  This is in preparation for the next commit, which
will refactor the logic for flushing an individual instance into
a separate function so that it can be called in other contexts
besides the full `pool-flush!`.