(SERVER-325) Move agent into PoolContext
The `borrow` functions are going to end up needing access to a bit
more state from the service in order to do things like flush
when the maximum number of requests has been exceeded for an
instance.  It looks like the easiest way to achive this will
be to extend the function signatures to accept a `PoolContext`
in addition to the pool queue.

This commit moves the pool-agent from the service context into
the `PoolContext` in preparation for these signature changes
to the `borrow` fns.