(SERVER-1052) Refactor jruby instance config to single fn
Prior to this commit, due to some shortcomings of the JRuby API
we had some duplicate code for initializing a JRuby instance.

We were setting things like `CompileMode`, load paths, environment
variables, etc. in two different places; one for a `ScriptingContainer`
and one for a `JRubyInstanceConfig`.  These classes have the same method
signatures for the methods that we care about, but because they
unfortunately don't share an interface, we were repeating the code
in two separate places because it seemed like it would be weird to
rely on the dynamic nature of Clojure to allow us to consolidate the
code into a single function that would work on either type of object.

As part of SERVER-1052 we will be adding some additional logic to
the configuration code in question, and rather than duplicating even
more logic I decided to go ahead and use a prismatic schema to
express that our config logic will work on either of those two
JRuby classes, and then go ahead and DRY up the code into a single
function.