(SERVER-1274) Purge JRubyClassLoader jars at container termination
This commit uses a hacky approach to ensure that handles to jars
registered with a ScriptingContainer's associated JRubyClassLoader
are closed and purged from the JRuby JarCache around the time that the
ScriptingContainer instance is terminated.  This avoids a leak of file
handles that would otherwise occur for those jars when a
ScriptingContainer falls out of scope.