(MAINT) Add 'smoke', 'resmoke' rake tasks
This commit refactors the `bakeNbeak` task a bit, so that it's
easy/possible to run the smoke test portion of it without the
ezbake portion of it.

It also introduces a 'resmoke' task that you can run if you want
to run tests again on a preserved host that resulted from an
earlier 'smoke' run.